telegram/requirements.txt
python-dotenv
apprise
